New feature: ability to detect double weaving (2+ advices per method)
This is a first step into the direction of permitting one instance or class to
be registered with multiple weavers, so each weaver can advise another (set of)
method(s) in one class. This can only work if the is an n to m relationship
between weavers and instances/classes. I.e. then it can also happen that method
matchers intersect, e.g. one weaver matches 'any()' in class X and another
matches 'nameStartsWith(get)'. The result would be that MethodAspect would be
woven into all getter methods twice (or multiple times in other situations) and
for each woven around interceptor pair the dynamic lookup in the advice registry
would be executed, leading to each advice being executed n times.

As it would be difficult and inefficient to dynamically inspect the byte code
for each visited method for pre-existing woven interceptors, the idea is to
have a registry of woven methods, which is currently implemented in a simplistic
way as 'static final WovenMethodRegistry wovenMethodRegistry' in class Weaver.
Because each time a new Weaver gets registered, Java instrumentation starts the
whole chain of previously applied transformations again from scratch, this kind
of bookkeeping only works if wovenMethodRegistry gets cleared when weaving
restarts. This is done via a custom AgentBuilder.InstallationListener.Adapter
registering a 'onBeforeInstall' hook and clearing 'wovenMethodRegistry' upon
call.

TODO: implement the actual n to m relationship.

A minor change in this commit is that 'StatisInitialiser*' classes, related
method names and comments have been renamed to 'TypeInitialiser*' or similar.
ByteBuddy uses the same naming scheme, only in American English ("initializer")
while I prefer British English ("initialiser"). Sorry to be pedantic here. ;-)

In preparation of future work there now is an 'Aspect' base class for classes
'*AroundAspect'.