Create new variant of uber JAR artifacts with relocated ASM & ByteBuddy classes
The new artifacts are:
  - aspect-core-no-dep (original: aspect-core)
  - remove-final-agent-no-dep (original: remove-final-agent)

Why these new artifacts? They can be used in projects where ASM and/or ByteBuddy
are used in incompatible versions directly or by other tools such as PowerMock.
Sometimes in such projects it is enough to just dependency-manage to a version
compatible with all callers (usually the more recent one), but in theory there
could be problems with incompatible or deprecated APIs, old/new bugs or
whatever. Anyway, in that case or just to be on the safe side if a user does not
mind having the same classes loaded multiple times, these new artifacts can be
used safely because all ASM and ByteBuddy classes have been relocated as
follows:
  - org.objectweb.asm -> net.bytebuddy.jar.asm -> de.scrum_master.jar.asm
  - net.bytebuddy -> de.scrum_master.jar.bytebuddy