New features: global mock, final class mock (work in progress)
I did not manage to implement constructor instrumentation the way I wanted to
with ASM, but succeeded with Javassist. The code first having been in a separate
project, I sourced it in in order to make it part of this project. The agent +
transformer can be built separately, but the ability to create uninitialised
object instances only really makes sense in combination with mocking/stubbing
methods, i.e. in combination with the aspect framework. Thus, the library can
also be used from within AspectAgent (which is actually no longer a good name
because it has evolved into some kind of "uber agent" containing
  - aspects,
  - constructor mocking / global mocking,
  - remove final.

TODO: a lot! For instance:
  - Make the aspect agent a separate agent again
  - Create new module for uber agent
  - Make agent infrastructure more template-like, e.g. by creating a base
    class/module for agents
  - Enable all agents to run in 3 modes:
      * Slim JAR: dependencies on the class path
      * Shaded JAR: dependencies inside the JAR in original packages
      * Relocated, shaded JAR: dependencies inside the JAR, external packages
        relocated to de.scrum_master.jar
    Each agent starter should recognise the mode automatically and inject the
    JAR(s) into the bootstrap class loader on on demand without configuration
    effort by the user
  - Make global mock transformation reversible
  - Write test(s) demonstrating how to use (global/final) mocks + advices
    together for mocking (now there are tests for the separate features only)
  - Write actual [Global-]Mock/Spy/Stub wrapper classes making mocking easier
    and more high level than now
  - Improve things like thread safety, multi class loader scenarios
  - Write documentation, both Javadoc and user documentation
  - and many more