Maven: centrally define agent/other JAR file locations relative to project root
Use maven.multiModuleProjectDirectory property as an absolute anchor point
instead of a plethora of module-specific relative paths, redefining the same
Maven properties and system properties, also possibly irritating IDEs.

P.S.: Even though it is officially undocumented and usage outside of Maven Core
is discouraged, this property is the best workaround for getting the project
root directory that I came up with so far.