Add the first integration test mocking constructor + methods of a final class
This is the fusion of the global mock and aspect advice features and just a
proof of concept for now. More streamlining and convenience classes are needed
and lots of open TODO items need to be cleared away. But for now it is working
quite nicely.

A first small bit of convenience has been added to class MethodAroundAdvice: a
static default instance MOCK with the default behaviour of returning null, 0 or
false for each static or instance method. This needs to be improved with regard
to stubbing a user selected set of methods, but for that the Weaver class needs
to support multiple advice pairs per class first, maybe via some kind of builder
API.