Introduce @Category(SeparateJVM) in order to optimise test JVM usage.
By default, re-use forked JVMs in Maven. Before the default was to use one JVM
per test class. This is now optional via test class annotation
@Category(SeparateJVM.class). For both Surefire and Failsafe there are now two
executions: One runs in re-use fork mode, the other one runs all classes of the
'SeparateJVM' category in its own JVM. There are still only a few tests
scattered across Maven modules, i.e. we do not save a lot of resources here yet,
but it is a good thing for the future and also a good example for users how to
do this in Maven and JUnit.

Unfortunately, this change also required to put JUnit into the project root POM
as a dependency, because otherwise otherwise Surefire/Failsafe groups and
excludedGroups would not work. This will not be fixed, see also
https://issues.apache.org/jira/browse/SUREFIRE-1266.

I also introduced a new package 'dev.sarek.agent.test' in the "test commons"
module, which has the additional advantage that all test utility classes which
should be exempt from byte code transformation are on the transformers' black
lists by default now.