Simplify Weaver constructor by removing Aspect.AdviceType parameter
The constructor can determine the around advice type by means of the new helper
method Aspect.forAdvice. So now there is only one Weaver constructor instead of
the previous four. The only small disadvantage is that when creating a type
initialiser weaver, we now have to pass a methodMatcher argument which then will
be ignored, so usually we would use null.