New feature: multiple advices of the same type per class/instance
This is now possible due to the usage of new class BiMultiMap in
Aspect.adviceRegistry. BiMultiMap is a bidirectional map implementing an n-to-m
relationship between keys and values which can also be used for reverse lookups.
BiMultiMap also synchronises add/remove methods.

Another improvement is that now for instance advices, a registered class is no
longer returned as a fall-back result as long as there are still non-Class
instances are registered for the same advice. Only if there are no registered
instances, a registered class will be returned, effectively making the advice
global for all instances.

If a user wants to build a special type of aspect or mock which only applies to
static methods, constructors or type initialisers, she simply can do one of the
following:

  a) Register no advices for instance methods.
  b) If there are instance method advices registered and global instance method
     mocking advising is unwanted, the user can just register a dummy instance
     which is not used elsewhere in the code.