Implement error handling + cleanup for errors in Weaver constructor
Now if there are any errors while adding targets from the constructor or
registering the transformer, an error handler calls 'unregisterTransformer()'.
This has the effect of removing all targets and subsequently also unregistering
the transformer. This call is again wrapped in a try-catch block in order to
ignore errors happening during clean-up, most prominently trying to unregister a
transformer even though none was registered yet because an error happened before
there was a chance to do so.