TypeInitialiserAspect: resolve class from system class loader
When TypeInitialiserAspect is on the bootstrap classpath, trying to resolve an
application class in order to access the advice for that class will fail. As the
type is not initialised yet, i.e. just being loaded, trying to load same type
in order to advise its type initialiser is kind of a hen vs. egg problem, so we
need to be careful here. We thus try to resolve the class from the system class
loader, but without initialising it yet, via:

Class.forName(className, false, ClassLoader.getSystemClassLoader())

I hope I got this right and there are no ugly corner cases.