Introduce Sarek multi agent (aspect agent is no longer a multi agent)
The new modules 'sarek' and 'sarek-agent' are now used and 'sarek-all.jar' added
as a test dependency rather than a multitude of other agent and core JARs, most
prominently the former multi agent 'aspect-core.jar' which was never meant to be
a multi agent in the first place.

Module 'mock' was moved to 'sarek-mock' and its tests to the no-agent test
module. TODO: check if the tests can be moved back.

More tests have been changed to run in their own JVMs due to problems without
that step.

The no-agent test POM was prepared to run with Sarek agent on the bootstrap
class path via command line, but currently this option is inactive. If it would
be activated, several expected exception for unmockable classes would then no
longer occur, requiring the tests to be frefactored.