Björn Kautler's constructor mock idea only works with '-noverify'
I implemented the idea, but modern JVMs, especially HotSpot, reject byte code
using
  try { super(); }
  catch (Exception e) { doSomething(); throw e; }
even though the Java byte code specification allows it. If I either just use
  super();
  doSomething();
or run the JVM with '-noverify', it works, but the former approach in not
transactional (fails if errors occur in super constructor) and the latter ugly.
Nobody wants to run all tests with '-noverify'.

So I am committing this in order to document (and maybe tag) it for reference.
Who knows if there might be a chance to implement something like this in the
future?