New class GlobalInstance<T> marks global instance mocking mode
Thanks to GlobalInstance<T> we can now differentiate between between global
class-based mocks (constructor and static method mocking) on the one hand and
global instance mocking (if no target class instances are registered) on the
other hand. There are convenience builder methods for this too and some tests
have been adjusted or extended in order to reflect the new capability. Better
test coverage and also Javadoc are still needed, but it seems to be working for
now.

There also is a new test MockIT in module agent-test-use-agent, which is almost
a copy of MockWithBootstrapInjectionTest, but because of the agent of course
does not need to manually append an agent JAR to the bootstrap classpath because
the Sarek Java agent already does that. A partially commented out test with
class-loading problems in runtime bootstrap injection mode now also works in
this mode, showing the difference between the two modes.