Improve hashCode/equals handling when using MockFactory API
Writing a simple test for hashCode stub generation uncovered a rather complex
problem I do not want to go into too deep here, but basically the root cause was
that targets in the advice registry were stored in data structures based on
object equality (based on equals/hashCode) instead of object identity (based on
objectA == objectB).

The necessary changes non-trivial, too:
  - The BiMultiMap now internally uses IdentityHashMap and IdentityArrayList.
    The latter class I copied from JDK 14.0.1 into this project and moved it to
    a Sarek package because the original Java module is not accessible by
    default.
  - I also had to add class EqualsAspect as an analog to HashCodeAspect.

A positive side effect is that now I see no more endlessly recursive method
call cascades which need to be detected by the other aspect classes, so maybe
the detection code in those 3 classes can be removed in the future. It no longer
seems to be relevant, but I am not sure enough yet, maybe there is a way to
reproduce those endless loops which lead to stack overflows and consequently
kill the JVM. So for now: better safe than sorry.