Commit previously moved files from modules sarek-aspect*
I don't know why IntelliJ IDEA messed this up, but the moved files now have to
be committed as new files because they were listed as untracked. I did not
notice because all the local builds were running fine because the files existed
in the work directory. Now there is a loss of direct traceability because there
is no direct file history for the whole aspect stuff anymore. :-(