Simplify aspect module structure, removing parent and agent
Actually, there is no real point in having AspectAgent anymore, since the multi
agent functionality was moved into SarekAgent. A stand-alone aspect agent also
does not make much sense because no global transformer needs to be started for
aspects (the same is true for high-level mocks and probably also for the
constructor mock transformation). Whenever a user needs a Weaver, she will just
create one, either directly or via mock API.

The only point in having a stand-alone aspect agent would be catering to users
who wish to use the aspect API, but not constructor mocks or the unfinal agent
and at the same time wish to inject the API and its dependencies into the
bootstrap classloader so as to be able to advise bootstrap classes. But I guess
it is acceptable to tell them to just use the multi agent and use the features
from sarek.jar selectively. Let's just keep it simple and not create more
infrastructure than necessary.