Add Sarek JUnit 4 test runner capable of bootstrap-injecting agent JARs
SarekRunner in combination with @SarekRunnerConfig provides for runtime agent
detection and agent configuration. The runner tries to attach one agent JAR per
JVM before any tests start running. Ideally, this enables the same behaviour as
if the JVM was started with the corresponding '-javaagent' argument. I adjusted
all tests in module 'sarek-test-no-agent' in order to synchronise their content
with the Java agent tests, and it looks as if everything working in the latter
also works in the former. :)

The downside is 95% test code duplication (gonna deal with it later), the main
differences are the new @RunWith(SarekRunner.class) and @SarekRunnerConfig
annotations.