Factor out hot attachment into new class AgentAttacher + add Spock extension
AgentAttacher is now used by both the JUnit 4 runner (SarekRunner) and the new
Spock global extension (SarekSpockExtension) which works for Spock 1.x and 2.x
without any changes. This is different from the JUnit 4 runner which will only
run in JUni5 5 Jupiter in compatibility mode. I will add a JUnit 5 extension
also utilising the AgentAttacher infrastructure soon.

Former annotation SarekRunnerConfig has been renamed to SarekRunnerDelegate and
only contains one value for the delegate class now. The rest of the
configuration lives in the new agent attacher package 'sarek-attach' now and is
based on three layers of Java properties in this order of precedence:
  1. system properties
  2. config file
  3. default values