ConstructorMockRegistry.isMockUnderConstruction now takes a Class parameter
This is a preparation for indexing constructor mocks by class in the future, no
longer by class name. It would allow to differentiate between classes with
identical names in different class loaders. For now the parameter is just
printed, not used otherwise. Both the Javassist and the ASM versions generate
the new call correctly.