Re-introduce agent auto detection which was accidentally removed in @c5fd9e1
When creating the default configuration, the agent attacher now checks for all
existing agent types if the corresponding JAR is found on the class path. This
is done via marker files with distinct magic names inside each JAR.

If an agent is detected, its symbolic name will be set as the default value of
system property 'dev.sarek.agent.type', otherwise the default is "".